Mongo

+++ ProductApplication/Models/Product.cs
@@ -2,7 +2,8 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
-
+using MongoDB.Bson;
+using MongoDB.Bson.Serialization.Attributes;
 namespace ProductApplication.Models
 {
      
@@ -12,11 +13,9 @@ namespace ProductApplication.Models
        
         public string  Name { get; set; }
         public decimal Price { get; set; }
-        public long ProductId { get; set; }
-        
+       
+        public int ProductId { get; set; }
         public Manufacturer ManufacturerDetails { get; set; }
-
         public int ProductInStock { get; set; }
-
     }
 }
